requirements.txt
numpy==1.18.4
xmltodict==0.12.0
pandas==1.0.3
redis==3.5.3
tqdm==4.46.0
pyarrow==0.17.1
scipy==1.4.1
dash==1.14.0
# Flask==1.1.2
Flask==2.3.2
plotly==4.6.0
dash_core_components==1.10.2
dash_extensions==0.0.26
py_vollib==1.0.1
SQLAlchemy==1.3.16
requests==2.22.0
mibian==0.1.3
spacy==2.2.4
dash_table==4.9.0
psycopg2==2.8.6
pytz==2020.1
pytextrank==2.0.2
selenium==3.141.0
dash_html_components==1.0.3
dash-core-components==1.10.2
#dashapp.egg==info
pandas_datareader==0.8.1
ipdb==0.13.2
python_dateutil==2.8.1
